requirements.txt
PyJWT==2.8.0
PyGithub==2.1.1
python-dotenv==1.0.0
uvicorn==0.27.0
fastapi==0.109.2
gunicorn==21.2.0
python-multipart==0.0.6